Add files via upload
Project – Build an iOS App (in Objective-C) with chrome and bottom tabs, mimicking the look and feel of Amazon App. With a top search bar and a link bar which scrolls horizontally, with more links. The Webview should load amazon.com. The tab bar has two tab items. When the webview scrolls, do the animation to hide the link bar and bottom tabs . With scroll up, hide animation. Scroll down, show both again.